refactor(workspace): host entity dialogs at the app shell
The create/edit and delete dialogs were rendered inside WorkspaceSidebar,
which unmounts when the sidebar is collapsed. Any surface that opened one
through useUiStore would then fail silently.

Move them into a dedicated workspace-dialogs module mounted next to
SettingsDialog, so the ui-store switches work regardless of the sidebar.
This also brings workspace-sidebar back under the module size guidance.
